requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
abseil-cpp=20200225.2=he1b5a44_2
argon2-cffi=20.1.0=py37h8f50634_1
arrow-cpp=1.0.1=py37h880e8b3_13_cpu
async_generator=1.10=py_0
attrs=20.2.0=pyh9f0ad1d_0
aws-c-common=0.4.59=he1b5a44_0
aws-c-event-stream=0.1.6=h84e28f3_5
aws-checksums=0.1.9=he252421_2
aws-sdk-cpp=1.8.59=h9b98462_1
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
bleach=3.2.1=pyh9f0ad1d_0
brotli=1.0.9=he1b5a44_2
brotlipy=0.7.0=py37h8f50634_1000
bzip2=1.0.8=h516909a_3
c-ares=1.16.1=h516909a_3
ca-certificates=2020.6.20=hecda079_0
certifi=2020.6.20=py37he5f6b98_2
cffi=1.14.3=py37h2b28604_0
chardet=3.0.4=py37he5f6b98_1008
click=7.1.2=pyh9f0ad1d_0
cryptography=3.1.1=py37hb09aad4_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py37hc8dfbb8_1001
findspark=1.3.0=py_1
gflags=2.2.2=he1b5a44_1004
glog=0.4.0=h49b9bf7_3
grpc-cpp=1.32.0=h7997a97_1
idna=2.10=pyh9f0ad1d_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
ipykernel=5.3.4=py37h43977f1_0
ipython=5.8.0=py37_1
ipython_genutils=0.2.0=py_1
jinja2=2.11.2=pyh9f0ad1d_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=py37hc8dfbb8_1
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py37hc8dfbb8_1
jupyterlab=2.2.8=py_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=1.2.0=py_0
krb5=1.17.1=hfafb76e_3
ld_impl_linux-64=2.35=h769bd43_9
libblas=3.8.0=17_openblas
libcblas=3.8.0=17_openblas
libcurl=7.71.1=hcdd3856_8
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=hcdb4288_2
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.3.0=h5dbcf3e_17
libgfortran-ng=7.5.0=hae1eefd_17
libgfortran4=7.5.0=hae1eefd_17
libgomp=9.3.0=h5dbcf3e_17
liblapack=3.8.0=17_openblas
libnghttp2=1.41.0=h8cfc5f6_2
libopenblas=0.3.10=pthreads_hb3c22a3_4
libprotobuf=3.13.0.1=h8b12597_0
libsodium=1.0.18=h516909a_1
libssh2=1.9.0=hab1572f_5
libstdcxx-ng=9.3.0=h2ae2ef3_17
libthrift=0.13.0=h5aa387f_6
libutf8proc=2.5.0=h516909a_2
lz4-c=1.9.2=he1b5a44_3
markupsafe=1.1.1=py37hb5d75c8_2
mistune=0.8.4=py37h8f50634_1001
nbclient=0.5.0=py_0
nbconvert=6.0.7=py37hc8dfbb8_0
nbformat=5.0.7=py_0
ncurses=6.2=he1b5a44_1
nest-asyncio=1.4.1=py_0
notebook=6.1.4=py37hc8dfbb8_0
numpy=1.19.1=py37h7ea13bd_2
openssl=1.1.1h=h516909a_0
orc=1.6.5=hd3605a7_0
packaging=20.4=pyh9f0ad1d_0
pandas=1.1.3=py37h3340039_0
pandoc=2.10.1=h516909a_0
pandocfilters=1.4.2=py_1
parquet-cpp=1.5.1=2
pexpect=4.8.0=py37hc8dfbb8_1
pickleshare=0.7.5=py37hc8dfbb8_1001
pip=20.2.3=py_0
prometheus_client=0.8.0=pyh9f0ad1d_0
prompt_toolkit=1.0.15=py_1
ptyprocess=0.6.0=py_1001
py4j=0.10.9=pyh9f0ad1d_0
pyarrow=1.0.1=py37hcae2be3_13_cpu
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.17.3=py37h8f50634_1
pysocks=1.7.1=py37hc8dfbb8_1
pyspark=3.0.1=pyh9f0ad1d_0
pyspark-stubs=3.0.0.post1=pypi_0
python=3.7.8=h6f2ec95_1_cpython
python-dateutil=2.8.1=py_0
python-dotenv=0.14.0=pyh9f0ad1d_0
python_abi=3.7=1_cp37m
pytz=2020.1=pyh9f0ad1d_0
pyzmq=19.0.2=py37hac76be4_1
re2=2020.10.01=he1b5a44_0
readline=8.0=he28a2e2_2
requests=2.24.0=pyh9f0ad1d_0
send2trash=1.5.0=py_0
setuptools=49.6.0=py37he5f6b98_2
simplegeneric=0.8.1=py_1
six=1.15.0=pyh9f0ad1d_0
snappy=1.1.8=he1b5a44_3
sqlite=3.33.0=h4cf870e_1
terminado=0.9.1=py37hc8dfbb8_0
testpath=0.4.4=py_0
tk=8.6.10=hed695b0_1
tornado=6.0.4=py37h8f50634_1
traitlets=5.0.4=py_1
urllib3=1.25.10=py_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.35.1=pyh9f0ad1d_0
xz=5.2.5=h516909a_1
zeromq=4.3.3=he1b5a44_2
zipp=3.3.0=py_0
zlib=1.2.11=h516909a_1009
zstd=1.4.5=h6597ccf_2
